feat: error dialog for boundary
